added WXS download for geuvadis, and made small tweaks to other notebooks

+++ scripts/run_vk_count_fig4_wxs.py
@@ -2,6 +2,7 @@ import os
 import subprocess
 import json
 import concurrent.futures
+import pandas as pd
 import varseek as vk
 
 # more on the datasets:
@@ -9,7 +10,8 @@ import varseek as vk
         # sequencing data (ENA): https://www.ebi.ac.uk/ena/browser/view/PRJNA523380
         # paper: https://www.nature.com/articles/s41586-019-1186-3
     # Geuvadis: 
-        # sequencing data (ENA): https://www.ebi.ac.uk/ena/browser/view/PRJEB3366
+        # RNA-seq sequencing data (ENA): https://www.ebi.ac.uk/ena/browser/view/PRJEB3366
+        # WXS sequencing data (IGSR): https://www.internationalgenome.org/data-portal/sample
         # paper: https://www.nature.com/articles/nature12531
 
 script_dir = os.path.dirname(os.path.abspath(__file__))
@@ -18,13 +20,14 @@ reference_out_dir = os.path.join(data_dir, "reference")
 
 # Sequencing database parameters
 sequencing_data_base = "geuvadis"  # "ccle" (Fig3) or "geuvadis" (Fig4) - ccle is ~11TB, geuvadis is ~2.3TB
-data_to_use = "rnaseq"  # only used when sequencing_data_base == "ccle"  # options: "rnaseq", "wxs", "wxs_with_corresponding_rnaseq_sample", "rnaseq_and_wxs", "wxs_and_rnaseq"
 number_of_threads_total = 32  # if too high (e.g., 64), then will not be able to download successfully (server error) - 8 seems like the sweet spot
 number_of_threads_per_varseek_count_task = 32
 max_retries = 5
-download_only = False
+download_only = True
 delete_fastq_files = False
 sequencing_data_out_base = os.path.join(data_dir, f"{sequencing_data_base}_data_base")
+geuvadis_rna_json_file = os.path.join(sequencing_data_out_base, f"{sequencing_data_base}_metadata.json")
+metadata_tsv_file = os.path.join(sequencing_data_out_base, "igsr_GRCh38.tsv")  # can be found at https://www.internationalgenome.org/data-portal/data-collection/grch38
 
 # reference parameters
 vk_ref_out = os.path.join(data_dir, "vk_ref_out")
@@ -76,58 +79,29 @@ if qc_against_gene_matrix and (not os.path.exists(reference_genome_index) or not
     reference_genome_f1 = os.path.join(reference_out_dir, "ensembl_grch37_release93", "f1.fasta")
     subprocess.run(["kb", "ref", "-t", str(number_of_threads_total), "-i", reference_genome_index, "-g", reference_genome_t2g, "-f1", reference_genome_f1, reference_genome_fasta, reference_genome_gtf], check=True)
 
-#* CCLE/Geuvadis
-ccle_ena_project = "PRJNA523380"
-geuvadis_ena_project = "PRJEB3366"
-if sequencing_data_base == "ccle":
-    ena_project = ccle_ena_project
-elif sequencing_data_base == "geuvadis":
-    ena_project = geuvadis_ena_project
-else:
-    raise ValueError("ccle or geuvadis")
+if not os.path.exists(metadata_tsv_file):
+    raise FileNotFoundError(f"Metadata TSV file not found: {metadata_tsv_file}. Please download it from https://www.internationalgenome.org/data-portal/data-collection/grch38 --> 'Available Data' --> 'Download the list' and place it in this path.")
 
-json_url = f"https://www.ebi.ac.uk/ena/portal/api/filereport?accession={ena_project}&result=read_run&fields=study_accession,sample_accession,experiment_accession,run_accession,scientific_name,library_strategy,experiment_title,experiment_alias,fastq_bytes,fastq_ftp,sra_ftp,sample_title&format=json&download=true&limit=0"
-
-# metadata json
-os.makedirs(sequencing_data_out_base, exist_ok=True)
-json_path = os.path.join(sequencing_data_out_base, f"{sequencing_data_base}_metadata.json")
-if not os.path.exists(json_path):
-    sequencing_metadata_download_command = ["wget", "-q", "-O", json_path, json_url]
+if not os.path.exists(geuvadis_rna_json_file):
+    json_url = "https://www.ebi.ac.uk/ena/portal/api/filereport?accession=PRJEB3366&result=read_run&fields=study_accession,sample_accession,experiment_accession,run_accession,scientific_name,library_strategy,experiment_title,experiment_alias,fastq_bytes,fastq_ftp,sra_ftp,sample_title&format=json&download=true&limit=0"
+    sequencing_metadata_download_command = ["wget", "-q", "-O", geuvadis_rna_json_file, json_url]
     subprocess.run(sequencing_metadata_download_command, check=True)
 
-    if sequencing_data_base == "ccle":
-        try:
-            json_updated_path_tmp = os.path.join(sequencing_data_out_base, "ccle_metadata_updated_tmp.json")
-            update_ccle_metadata_script = os.path.join(script_dir, "update_ccle_metadata.R")
-            subprocess.run(["Rscript", update_ccle_metadata_script, json_path, json_updated_path_tmp], check=True)  # TODO: will try to install dependencies in conda environment - unsure if this works
-            os.rename(json_updated_path_tmp, json_path)
-        except Exception as e:
-            print("Error in updating CCLE metadata:", e)
-
 # Loop through json file and download fastqs
-with open(json_path, 'r', encoding="utf-8") as file:
+with open(geuvadis_rna_json_file, 'r', encoding="utf-8") as file:
     data = json.load(file)
 
-if sequencing_data_base == "ccle":
-    rnaseq_data = [study for study in data if study['library_strategy'] == 'RNA-Seq']
-    wxs_data = [study for study in data if study['library_strategy'] == 'WXS']
-    rnaseq_study_accessions = {study['study_accession'] for study in rnaseq_data}
-    wxs_data_with_corresponding_rnaseq_sample = [study for study in wxs_data if study['sample_accession'] in rnaseq_study_accessions]
+rna_df = pd.DataFrame(data)
+sample_titles_set = set(rna_df['sample_title'].tolist())
 
-    if data_to_use.lower() == "rnaseq":
-        data_list_to_run = rnaseq_data
-    elif data_to_use.lower() == "wxs":
-        data_list_to_run = wxs_data
-    elif data_to_use.lower() == "wxs_with_corresponding_rnaseq_sample":
-        data_list_to_run = wxs_data_with_corresponding_rnaseq_sample
-    elif data_to_use.lower() == "rnaseq_and_wxs" or data_to_use.lower() == "wxs_and_rnaseq":
-        data_list_to_run = rnaseq_data + wxs_data
-    else:
-        raise ValueError("data_to_use must be one of 'rnaseq', 'wxs', 'wxs_with_corresponding_rnaseq_sample', 'rnaseq_and_wxs', or 'wxs_and_rnaseq'")
-elif sequencing_data_base == "geuvadis":
-    data_list_to_run = data
-
-number_of_items = len(data_list_to_run)
+wxs_df = pd.read_csv("/Users/joeyrich/Downloads/geuvadis_wxs.tsv", sep="\t", low_memory=False, usecols=["Sample", "Data type", "Analysis group", "url"])
+wxs_df = wxs_df[
+    (wxs_df["Sample"].isin(sample_titles_set)) &
+    (wxs_df["Data type"] == "sequence") &
+    (wxs_df["Analysis group"] == "Exome")
+]
+wxs_df = wxs_df.drop_duplicates(subset=["Sample"])
+number_of_items = len(wxs_df)
 
 if download_only:
     number_of_threads_per_varseek_count_task = 1
@@ -139,7 +113,7 @@ if save_vcf and not os.path.exists(vcf_data_csv):  # alternatively, I can do thi
     vk.utils.add_vcf_info_to_cosmic_tsv(cosmic_tsv=cosmic_tsv, reference_genome_fasta=cosmic_reference_genome_fasta, cosmic_df_out=vcf_data_csv, sequences=sequences, cosmic_version=101)
 
 def download_sequencing_total(
-    record,
+    row,
     vcrs_index,
     vcrs_t2g,
     sequencing_data_out_base = ".",
@@ -155,12 +129,8 @@ def download_sequencing_total(
     vcf_data_csv=None,
     number_of_threads_per_varseek_count_task=2,
 ):
-    experiment_alias = record.get('experiment_alias')
-    experiment_alias_underscores_only = experiment_alias.replace("-", "_")
-    sample = experiment_alias_underscores_only  # f"{experiment_alias_underscores_only}___{sample_accession}___{experiment_accession}___{run_accession}"
-
-    fastq_ftp = record.get('fastq_ftp')
-    fastq_links = fastq_ftp.split(';')
+    sample = row['Sample']
+    link = row['url']
 
     sample_out_folder = os.path.join(sequencing_data_out_base, sample)
     os.makedirs(sample_out_folder, exist_ok=True)
@@ -169,35 +139,33 @@ def download_sequencing_total(
     fastq_files = list()
 
     all_files_downloaded = True
-    for link in fastq_links:
-        rnaseq_fastq_file = os.path.join(sample_out_folder, os.path.basename(link))
-        tmp_fastq_file = os.path.join(sample_out_folder, "tmp_" + os.path.basename(link))  # Temporary file, just so I know files that have not finished downloading
-        fastq_files.extend([rnaseq_fastq_file, tmp_fastq_file])
+    wxs_fastq_file = os.path.join(sample_out_folder, os.path.basename(link))
+    tmp_fastq_file = os.path.join(sample_out_folder, "tmp_" + os.path.basename(link))  # Temporary file, just so I know files that have not finished downloading
+    fastq_files.extend([wxs_fastq_file, tmp_fastq_file])
 
-        if not os.path.exists(rnaseq_fastq_file):  # just here while I keep files permanently for debugging
-            all_files_downloaded = False
-            
-            if os.path.exists(tmp_fastq_file):  # If an old tmp file exists, delete it to ensure a clean restart
-                print(f"Removing incomplete file: {tmp_fastq_file}")
-                os.remove(tmp_fastq_file)
-            
-            print(f"Downloading {link} to {rnaseq_fastq_file}")
+    if not os.path.exists(wxs_fastq_file):  # just here while I keep files permanently for debugging
+        all_files_downloaded = False
+        
+        if os.path.exists(tmp_fastq_file):  # If an old tmp file exists, delete it to ensure a clean restart
+            print(f"Removing incomplete file: {tmp_fastq_file}")
+            os.remove(tmp_fastq_file)
+        
+        print(f"Downloading {link} to {wxs_fastq_file}")
 
-            if not link.startswith(('ftp://', 'http://')):
-                link = 'ftp://' + link
+        if not link.startswith(('ftp://', 'http://')):
+            link = 'ftp://' + link
 
-            # download_command = f"curl --connect-timeout 60 --speed-time 30 --speed-limit 10000 -o {rnaseq_fastq_file} {link}"
-            download_command = f"wget -c --tries={max_retries} --retry-connrefused -O {tmp_fastq_file} {link}"  # --limit-rate=1m  (or some other rate)
-            # download_command = f"aria2c -x 16 -d {sample_out_folder} -o {os.path.basename(link)} -c {link}"
+        # download_command = f"curl --connect-timeout 60 --speed-time 30 --speed-limit 10000 -o {wxs_fastq_file} {link}"
+        download_command = f"wget -c --tries={max_retries} --retry-connrefused -O {tmp_fastq_file} {link}"  # --limit-rate=1m  (or some other rate)
+        # download_command = f"aria2c -x 16 -d {sample_out_folder} -o {os.path.basename(link)} -c {link}"
 
-            try:
-                result = subprocess.run(download_command, shell=True, check=True)
-                os.rename(tmp_fastq_file, rnaseq_fastq_file)  # If successful, rename to final filename
-            except subprocess.CalledProcessError as e:
-                print(f"Error downloading {link} to {rnaseq_fastq_file}")
-                print(e)
-                failed_downloads.append(rnaseq_fastq_file)
-                continue
+        try:
+            result = subprocess.run(download_command, shell=True, check=True)
+            os.rename(tmp_fastq_file, wxs_fastq_file)  # If successful, rename to final filename
+        except subprocess.CalledProcessError as e:
+            print(f"Error downloading {link} to {wxs_fastq_file}")
+            print(e)
+            failed_downloads.append(wxs_fastq_file)
     
     # count number of dirs in sequencing_data_out_base
     if not all_files_downloaded:
@@ -238,10 +206,8 @@ def download_sequencing_total(
 
     if delete_fastq_files:
         for fastq_file in fastq_files:
-            os.remove(fastq_file)
-
-    import sys  #!!! erase
-    sys.exit()  #!!! erase
+            if os.path.isfile(fastq_file):
+                os.remove(fastq_file)
 
 
 number_of_tasks = number_of_threads_total / number_of_threads_per_varseek_count_task
@@ -249,7 +215,7 @@ with concurrent.futures.ThreadPoolExecutor(max_workers=number_of_tasks) as execu
     futures = [
         executor.submit(
             download_sequencing_total,
-            record=record,
+            row=row,
             vcrs_index=vcrs_index,
             vcrs_t2g=vcrs_t2g,
             sequencing_data_out_base=sequencing_data_out_base,
@@ -265,7 +231,7 @@ with concurrent.futures.ThreadPoolExecutor(max_workers=number_of_tasks) as execu
             vcf_data_csv=vcf_data_csv,
             number_of_threads_per_varseek_count_task=number_of_threads_per_varseek_count_task,
         )
-        for record in data_list_to_run
+        for _, row in wxs_df.iterrows()
     ]
 
     concurrent.futures.wait(futures)
